+++ core/jvm/src/Sink.kt
@@ -34,9 +34,9 @@ public actual sealed interface Sink : RawSink, WritableByteChannel {
 
   public actual fun write(source: RawSource, byteCount: Long): Sink
 
-  public actual fun writeByte(byte: Int): Sink
+  public actual fun writeByte(byte: Byte): Sink
 
-  public actual fun writeShort(short: Int): Sink
+  public actual fun writeShort(short: Short): Sink
 
   public actual fun writeInt(int: Int): Sink
 
@@ -77,7 +77,7 @@ public fun Sink.outputStream(): OutputStream {
   return object : OutputStream() {
     override fun write(b: Int) {
       if (!isOpen) throw IOException("closed")
-      buffer.writeByte(b.toByte().toInt())
+      buffer.writeByte(b.toByte())
       emitCompleteSegments()
     }
